extracted field.py module, add test_request_objects
